Simplify Roadmap Components
